// src/model/Hobby.ts
export class Hobby {
    Travel_Service_Id: number;
    Travel_Service_Type: string;
    User_Created: number;
    Travel_Service_Icon: string;
    Travel_Service_Location: string;
    Travel_Service_Name: string;
    Travel_Service_Description: string;
    Service_Level: string;
    Language_Id: string;
    NAME: string;
    ID: number;
    CHECKED: boolean;
}